pyproject.toml
[tool.poetry]
name = "com-build-api"
version = "0.1.0"
description = ""
authors = ["xiomara huamani <101735835+XiomaraHuamani@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
httpx = "^0.28.1"
pydantic = {extras = ["email"], version = "^2.11.2"}
alembic = "^1.15.2"
pydantic-settings = "^2.8.1"
db = "^0.1.1"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
email-validator = "^2.2.0"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
passlib = "^1.7.4"
bcrypt = "3.2.2"
uvicorn = "^0.34.0"
fastapi = "^0.115.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
